BUG: Use ptrdiff_t instead of int for pointer offsets in CPU projectors
Fixes #979, i.e., segmentation faults when projecting large volumes
causing int overflow.

+++ include/rtkJosephBackAttenuatedProjectionImageFilter.h
@@ -55,7 +55,7 @@ public:
   }
 
   TOutput
-  operator()(const double stepLengthInVoxel, const TCoordinateType weight, const TInput * p, const int i)
+  operator()(const double stepLengthInVoxel, const TCoordinateType weight, const TInput * p, const ptrdiff_t i)
   {
     const double w = weight * stepLengthInVoxel;
